Match package and directory names
As described in the README and Go docs on "How to write Go code"
(https://go.dev/doc/code), a Go package is files in the same directory
that are compiled together, whereas a Go module is a collection of
packages released together. Executable commands use `package main`, but
`package main` cannot be imported elsewhere, so it's not ideal to use
`package main` for all source code.

The code here was a mix of `package main` and specific package names.
This commit will update the code so that the package and directory names
match and so that the examples don't use `package main`.