Remove tour directory
This directory was from A Tour of Go (https://go.dev/tour/welcome/1).
The other directories are from Learn Go with Tests
(https://github.com/quii/learn-go-with-tests).
The tour code is not needed and will be removed. This way, the code in
the repo aligns more closely with Learn Go with Tests.